+++ test/bones_api_entity_adapter_memory_test.dart
@@ -29,6 +29,5 @@ Future<bool> _runTest(bool useReflection) => runAdapterTests(
       ),
       '"',
       'int',
-      contains('ok'),
       entityByReflection: useReflection,
     );